Added dark mode and changed the color scheme of the login forum
Changed the color scheme of the background of the page as well as adding a system-ui for font family to take into any user's font they may have customized to their browser. added HTML scroll heavior to smooth and changed color scheme as well as changing the the button size and adding a shadow effect to it.